+++ larcore/Geometry/CMakeLists.txt
@@ -1,12 +1,13 @@
 
 cet_build_plugin(AuxDetGeometry art::service
-                 LIBRARIES
+                 LIBRARIES PUBLIC
                            larcorealg::Geometry
                            larcoreobj::SummaryData
                            art::Framework_Services_Registry
-                           art::Framework_Principal
                            art::Framework_Core
                            art::Persistency_Provenance
+                         PRIVATE
+                           art::Framework_Principal
                            messagefacility::MF_MessageLogger
                            ROOT::Core)
 cet_build_plugin(DumpChannelMap  art::module
@@ -28,7 +29,7 @@ cet_build_plugin(DumpGeometry  art::module
                           ROOT::Geom
                           ROOT::GenVector)
 cet_build_plugin(GeometryConfigurationWriter art::service
-                 LIBRARIES
+                 LIBRARIES REG
                            larcorealg::Geometry
                            larcoreobj::SummaryData
                            art::Framework_Services_Registry
@@ -38,13 +39,14 @@ cet_build_plugin(GeometryConfigurationWriter art::service
                            messagefacility::MF_MessageLogger
                            ROOT::Core)
 cet_build_plugin(Geometry art::service
-                 LIBRARIES
+                 LIBRARIES PUBLIC
                            larcorealg::Geometry
                            larcoreobj::SummaryData
                            art::Framework_Services_Registry
-                           art::Framework_Principal
                            art::Framework_Core
                            art::Persistency_Provenance
+                         PRIVATE
+                           art::Framework_Principal
                            messagefacility::MF_MessageLogger
                            ROOT::Core)
 cet_build_plugin(StandardGeometryHelper art::service
